Changed Resources Caravana
Try for fix the form bug and added auto remove on caravana delete.

+++ Back-End/src/main/java/me/caravanweb/resources/api/CaravanasResource.java
@@ -1,6 +1,5 @@
 package me.caravanweb.resources.api;
 
-import java.net.URI;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -13,12 +12,12 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-
 import me.caravanweb.profiles.Caravanas;
 import me.caravanweb.services.CaravanasService;
+import me.caravanweb.services.UsuarioService;
 
 @RestController
 @CrossOrigin
@@ -26,6 +25,8 @@ import me.caravanweb.services.CaravanasService;
 public class CaravanasResource {
 	@Autowired
 	private CaravanasService service;
+	@Autowired
+	private UsuarioService serviceu;
 	
 	@GetMapping
 	public ResponseEntity<List<Caravanas>> findAll() { 
@@ -46,7 +47,7 @@ public class CaravanasResource {
 	
 	@PostMapping
 	@ResponseStatus(HttpStatus.CREATED)
-	public ResponseEntity<String> create(@RequestBody Caravanas c){
+	public @ResponseBody ResponseEntity<String> create(@RequestBody Caravanas c){
 		String body =  service.add(c);
 		return ResponseEntity.ok().body(body);
 	}
@@ -54,12 +55,15 @@ public class CaravanasResource {
     @PostMapping(value = "/delete/{id}")
     @ResponseStatus(HttpStatus.CREATED)
     public ResponseEntity<String> delete(@PathVariable Integer id){
-    	Caravanas caravana = service.findById(id);
     	String body = "";
-		boolean status = service.delete(caravana);
-		if (status) {
-			body = "Deletado.";
-		}
+    	Caravanas caravana = service.findById(id);
+    	if (caravana != null) {
+    		body = serviceu.cleanCaravanaUsers(id);
+    		boolean status = service.delete(caravana);
+    		if (status) {
+    			body = "Deletado.";
+    		}
+    	}
 		return ResponseEntity.ok().body(body);    
 	}
